+++ FuturesTests/Future+MergeTests.swift
@@ -2,26 +2,12 @@ import Quick
 import Nimble
 import PiedPiper
 
-extension MutableCollectionType where Self.Index == Int {
-  mutating func shuffle() -> Self {
-    let numberOfElements = self.count
-    for iteration in 0..<(numberOfElements - 1) {
-      let swapIndex = Int(arc4random_uniform(UInt32(numberOfElements)))
-      if iteration != swapIndex {
-        swap(&self[iteration], &self[swapIndex])
-      }
-    }
-    
-    return self
-  }
-}
-
-class FutureSequenceReduceTests: QuickSpec {
+class FutureSequenceMergeTests: QuickSpec {
   override func spec() {
-    describe("Reducing a list of Futures") {
+    describe("Merging a list of Futures") {
       var promises: [Promise<Int>]!
-      var reducedFuture: Future<Int>!
-      var successValue: Int?
+      var mergedFuture: Future<[Int]>!
+      var successValue: [Int]?
       var failureValue: ErrorType?
       var wasCanceled: Bool!
       
@@ -38,11 +24,11 @@ class FutureSequenceReduceTests: QuickSpec {
         successValue = nil
         failureValue = nil
         
-        reducedFuture = promises
+        mergedFuture = promises
           .map { $0.future }
-          .reduce(5, combine: +)
+          .merge()
         
-        reducedFuture.onCompletion { result in
+        mergedFuture.onCompletion { result in
           switch result {
           case .Success(let value):
             successValue = value
@@ -62,7 +48,7 @@ class FutureSequenceReduceTests: QuickSpec {
           promises[1].fail(expectedError)
         }
         
-        it("should fail the reduced future") {
+        it("should fail the merged future") {
           expect(failureValue).notTo(beNil())
         }
         
@@ -70,11 +56,11 @@ class FutureSequenceReduceTests: QuickSpec {
           expect(failureValue as? TestError).to(equal(expectedError))
         }
         
-        it("should not cancel the reduced future") {
+        it("should not cancel the merged future") {
           expect(wasCanceled).to(beFalse())
         }
         
-        it("should not succeed the reduced future") {
+        it("should not succeed the merged future") {
           expect(successValue).to(beNil())
         }
       }
@@ -85,40 +71,39 @@ class FutureSequenceReduceTests: QuickSpec {
           promises[1].cancel()
         }
         
-        it("should not fail the reduced future") {
+        it("should not fail the merged future") {
           expect(failureValue).to(beNil())
         }
         
-        it("should cancel the reduced future") {
+        it("should cancel the merged future") {
           expect(wasCanceled).to(beTrue())
         }
         
-        it("should not succeed the reduced future") {
+        it("should not succeed the merged future") {
           expect(successValue).to(beNil())
         }
       }
       
       context("when all the original futures succeed") {
-        var expectedResult: Int!
+        var expectedResult: [Int]!
         
         context("when they succeed in the same order") {
           beforeEach {
-            expectedResult = 5
+            expectedResult = promises.enumerate().map { $0.index }
             promises.enumerate().forEach { (iteration, promise) in
               promise.succeed(iteration)
-              expectedResult = expectedResult + iteration
             }
           }
           
-          it("should not fail the reduced future") {
+          it("should not fail the merged future") {
             expect(failureValue).to(beNil())
           }
           
-          it("should not cancel the reduced future") {
+          it("should not cancel the merged future") {
             expect(wasCanceled).to(beFalse())
           }
           
-          it("should succeed the reduced future") {
+          it("should succeed the merged future") {
             expect(successValue).notTo(beNil())
           }
           
@@ -129,11 +114,11 @@ class FutureSequenceReduceTests: QuickSpec {
       }
     }
     
-    describe("Reducing a list of Futures, independently of the order they succeed") {
+    describe("Merging a list of Futures, independently of the order they succeed") {
       var promises: [Promise<String>]!
-      var reducedFuture: Future<String>!
-      var successValue: String?
-      var expectedResult: String!
+      var mergedFuture: Future<[String]>!
+      var successValue: [String]?
+      var expectedResult: [String]!
       
       beforeEach {
         promises = [
@@ -146,29 +131,28 @@ class FutureSequenceReduceTests: QuickSpec {
         
         successValue = nil
         
-        reducedFuture = promises
+        mergedFuture = promises
           .map { $0.future }
-          .reduce("BEGIN-", combine: +)
+          .merge()
         
-        reducedFuture.onSuccess {
+        mergedFuture.onSuccess {
           successValue = $0
         }
         
-        let sequenceOfIndexes = Array(0..<promises.count).map({ "\($0)" }).joinWithSeparator("")
-        expectedResult = "BEGIN-\(sequenceOfIndexes)"
+        expectedResult = Array(0..<promises.count).map { "\($0)" }
         
         var arrayOfIndexes = Array(promises.enumerate())
         
         repeat {
           arrayOfIndexes = arrayOfIndexes.shuffle()
         } while arrayOfIndexes.map({ $0.0 }) == Array(0..<promises.count)
-          
+        
         arrayOfIndexes.forEach { (originalIndex, promise) in
           promise.succeed("\(originalIndex)")
         }
       }
       
-      it("should succeed the reduced future") {
+      it("should succeed the merged future") {
         expect(successValue).notTo(beNil())
       }